Extract signers helpers
